# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  version: 1.0.0
  description: API для управления банковскими картами, переводами и администрированием пользователей.

servers:
  - url: http://localhost:8080
    description: Local environment

tags:
  - name: Auth
  - name: Cards
  - name: Transfers
  - name: Admin Cards
  - name: Admin Users

paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/cards:
    get:
      tags: [Cards]
      summary: Get user cards with search and pagination
      operationId: getUserCards
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: number
          schema:
            type: string
          required: false
          description: Search by card number substring
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Cards page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseCard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/cards/block:
    post:
      tags: [Cards]
      summary: Create block request
      operationId: blockCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockCardRequest'
      responses:
        '200':
          description: Block request created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/cards/{cardId}/balance:
    get:
      tags: [Cards]
      summary: Get card balance
      operationId: getCardBalance
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/transfers:
    post:
      tags: [Transfers]
      summary: Make transfer between own cards
      operationId: makeTransfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Transfers]
      summary: Get user transfers with pagination
      operationId: getUserTransfers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Transfers page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseTransfer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards:
    get:
      tags: [Admin Cards]
      summary: List all cards (admin)
      operationId: adminListCards
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Cards page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseAdminCard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Admin Cards]
      summary: Create card for user (admin)
      operationId: adminCreateCard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Card created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/{cardId}:
    get:
      tags: [Admin Cards]
      summary: Get card by id (admin)
      operationId: adminGetCard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Admin Cards]
      summary: Delete card (admin)
      operationId: adminDeleteCard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/{cardId}/block:
    post:
      tags: [Admin Cards]
      summary: Block card (admin)
      operationId: adminBlockCard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card blocked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/{cardId}/activate:
    post:
      tags: [Admin Cards]
      summary: Activate card (admin)
      operationId: adminActivateCard
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Card activated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/block-requests:
    get:
      tags: [Admin Cards]
      summary: List pending block requests (admin)
      operationId: adminListPendingBlockRequests
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Pending block requests page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseAdminCardBlockRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/block-requests/{requestId}/approve:
    post:
      tags: [Admin Cards]
      summary: Approve card block request (admin)
      operationId: adminApproveBlockRequest
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Block request approved
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/cards/block-requests/{requestId}/reject:
    post:
      tags: [Admin Cards]
      summary: Reject card block request (admin)
      operationId: adminRejectBlockRequest
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Block request rejected
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/users:
    get:
      tags: [Admin Users]
      summary: List users (admin)
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Users page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseAdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/users/{userId}:
    get:
      tags: [Admin Users]
      summary: Get user by id (admin)
      operationId: adminGetUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Admin Users]
      summary: Delete user (admin)
      operationId: adminDeleteUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/users/{userId}/active:
    patch:
      tags: [Admin Users]
      summary: Set user active flag (admin)
      operationId: adminSetUserActive
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserActiveRequest'
      responses:
        '204':
          description: User activity updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/users/{userId}/roles:
    put:
      tags: [Admin Users]
      summary: Set user roles (admin)
      operationId: adminSetUserRoles
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserRolesRequest'
      responses:
        '204':
          description: User roles updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Validation or business error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: secret123

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: secret123
        firstName:
          type: string
          example: Ivan
        lastName:
          type: string
          example: Ivanov

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        roles:
          type: array
          items:
            type: string
          example: [ROLE_USER]

    BlockCardRequest:
      type: object
      required: [cardId]
      properties:
        cardId:
          type: string
          format: uuid

    BalanceResponse:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        cardNumber:
          type: string
          example: '**** **** **** 5678'
        balance:
          type: integer
        status:
          type: string
          example: ACTIVE

    CardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
          example: '**** **** **** 5678'
        balance:
          type: integer
        owner:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CardStatusEnum'

    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: string
          format: uuid
        toCardId:
          type: string
          format: uuid
        amount:
          type: integer
          minimum: 1
          example: 100

    TransferCardInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        balance:
          type: integer
        number:
          type: string
          example: '**** **** **** 5678'

    TransferResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        balance:
          type: integer
        created:
          type: string
          format: date-time
        fromCard:
          $ref: '#/components/schemas/TransferCardInfo'
        fromCardNumber:
          type: string
          example: '**** **** **** 5678'
        toCard:
          $ref: '#/components/schemas/TransferCardInfo'
        toCardNumber:
          type: string
          example: '**** **** **** 9876'

    CreateCardRequest:
      type: object
      required: [userId, number, owner, expiryDate, balance]
      properties:
        userId:
          type: string
          format: uuid
        number:
          type: string
          example: '1234567890123456'
        owner:
          type: string
          example: Ivan Ivanov
        expiryDate:
          type: string
          format: date
        balance:
          type: integer
          minimum: 0

    AdminCardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        maskedNumber:
          type: string
          example: '**** **** **** 5678'
        owner:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CardStatusEnum'
        balance:
          type: integer
        userId:
          type: string
          format: uuid
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time

    AdminCardBlockRequestResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        cardId:
          type: string
          format: uuid
        maskedCardNumber:
          type: string
          example: '**** **** **** 5678'
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/CardBlockRequestStatusEnum'
        created:
          type: string
          format: date-time

    SetUserActiveRequest:
      type: object
      required: [active]
      properties:
        active:
          type: boolean

    SetUserRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          minItems: 1
          items:
            type: string
          example: [ADMIN]

    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        active:
          type: boolean
        roles:
          type: array
          items:
            type: string
          example: [ROLE_USER, ROLE_ADMIN]
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time

    PaginationItem:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer

    PaginationResponseCard:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        pagination:
          $ref: '#/components/schemas/PaginationItem'

    PaginationResponseTransfer:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransferResponse'
        pagination:
          $ref: '#/components/schemas/PaginationItem'

    PaginationResponseAdminCard:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminCardResponse'
        pagination:
          $ref: '#/components/schemas/PaginationItem'

    PaginationResponseAdminUser:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserResponse'
        pagination:
          $ref: '#/components/schemas/PaginationItem'

    PaginationResponseAdminCardBlockRequest:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminCardBlockRequestResponse'
        pagination:
          $ref: '#/components/schemas/PaginationItem'

    CardStatusEnum:
      type: string
      enum: [ACTIVE, BLOCKED, DEADLINE_EXPIRED]

    CardBlockRequestStatusEnum:
      type: string
      enum: [PENDING, APPROVED, REJECTED]

    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: Bad Request
        message:
          type: string
        path:
          type: string